Monitoraggio degli episodi, schermata Episodi, implementazione varie funzionalità schermate di dettaglio di film e serie TV, implementazione follow, implementazione funzionalità schermata di dettaglio utenti, visualizzazione delle liste

+++ WatchWise/Model/TVShow.swift
@@ -15,7 +15,7 @@ struct TVShow: Codable {
     let firstAirDate: String?
     let genres: [Genre]
     let homepage: String?
-    let id: Int
+    let id: Int64
     let inProduction: Bool?
     let languages: [String]?
     let lastAirDate: String?
